+++ resources/views/components/layouts/app.blade.php
@@ -1,5 +1,7 @@
+{{-- <x-layouts.app.header>
+    {{ $slot }}
+</x-layouts.app.header> --}}
+
 <x-layouts.app.sidebar>
-    <flux:main>
-        {{ $slot }}
-    </flux:main>
+    {{ $slot }}
 </x-layouts.app.sidebar>