+++ resources/views/components/layouts/auth/split.blade.php
@@ -1,15 +1,15 @@
 <!DOCTYPE html>
-<html lang="{{ str_replace('_', '-', app()->getLocale()) }}" class="dark">
+<html lang="{{ str_replace('_', '-', app()->getLocale()) }}">
     <head>
         @include('partials.head')
     </head>
-    <body class="min-h-screen bg-white antialiased dark:bg-linear-to-b dark:from-neutral-950 dark:to-neutral-900">
+    <body class="min-h-screen bg-base-200 antialiased">
         <div class="relative grid h-dvh flex-col items-center justify-center px-8 sm:px-0 lg:max-w-none lg:grid-cols-2 lg:px-0">
-            <div class="bg-muted relative hidden h-full flex-col p-10 text-white lg:flex dark:border-r dark:border-neutral-800">
-                <div class="absolute inset-0 bg-neutral-900"></div>
+            <div class="bg-base-300 relative hidden h-full flex-col p-10 text-base-content lg:flex dark:border-r dark:border-neutral-800">
+                <div class="absolute inset-0 xbg-neutral-900"></div>
                 <a href="{{ route('home') }}" class="relative z-20 flex items-center text-lg font-medium" wire:navigate>
                     <span class="flex h-10 w-10 items-center justify-center rounded-md">
-                        <x-app-logo-icon class="mr-2 h-7 fill-current text-white" />
+                        <x-app-logo-icon class="mr-2 h-7 fill-current text-base-content" />
                     </span>
                     {{ config('app.name', 'Laravel') }}
                 </a>
@@ -38,6 +38,5 @@
                 </div>
             </div>
         </div>
-        @fluxScripts
     </body>
 </html>